+++ Ficheiros/src/main/java/com/mycompany/ficheiros/Ficheiros.java
@@ -47,9 +47,9 @@ public class Ficheiros {
         }
 
         // Classificação de cada aluno
-        List<Integer> notasAlunos = new ArrayList<>();
+        List<Double> notasAlunos = new ArrayList<>();
         for (List<Integer> respostas : respostasAlunos) {
-            int nota = 0;
+            double nota = 0.0;
             for (int i = 0; i < respostas.size(); i++) { // Comparação das respostas dos aliunos com as corretas
                 if (respostas.get(i) == chaveRespostas.get(i)) {
                     nota += 1; // 1 ponto pela resposta certa
@@ -61,10 +61,10 @@ public class Ficheiros {
         }
 
         // Nota mais alta e nota mais baixa
-        int notaMax = notasAlunos.get(0); // notaMax inicializado com o 1 elemento da lista
-        int notaMin = notasAlunos.get(0); // notMin inicializado com o 1 elemento da lista
+        double notaMax = notasAlunos.get(0); // notaMax inicializado com o 1 elemento da lista
+        double notaMin = notasAlunos.get(0); // notMin inicializado com o 1 elemento da lista
         for (int i = 1; i < notasAlunos.size(); i++) { // Feito um loop para comparar a restante lista com as variáveis
-            int nota = notasAlunos.get(i);
+            double nota = notasAlunos.get(i);
             if (nota > notaMax) {
                 notaMax = nota; // Máxima nota
             }
@@ -76,14 +76,14 @@ public class Ficheiros {
         // Média das notas
         double somaNotas = 0;
         double mediaNotas = 0;
-        for (int nota : notasAlunos) {
+        for (double nota : notasAlunos) {
             somaNotas += nota;
         }
         mediaNotas = somaNotas / notasAlunos.size(); // Média final a dividir pelos 30 alunos
 
         // Número de positivas e respetiva percentagem
         int numPositivas = 0;
-        for (int nota : notasAlunos) {
+        for (double nota : notasAlunos) {
             if (nota >= 10) {
                 numPositivas++;
             }
